get historical prices and use rust_decimal

+++ src/rest/model/balance.rs
@@ -1,8 +1,10 @@
+use rust_decimal::Decimal;
+
 #[derive(Debug, serde::Serialize, serde::Deserialize)]
 pub struct Balance {
     pub coin: String,
-    pub free: f64,
-    pub total: f64,
-    pub spot_borrow: f64,
-    pub available_without_borrow: f64,
+    pub free: Decimal,
+    pub total: Decimal,
+    pub spot_borrow: Decimal,
+    pub available_without_borrow: Decimal,
 }
