+++ src/rest/model/market.rs
@@ -1,3 +1,4 @@
+use rust_decimal::Decimal;
 use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Serialize, Deserialize)]
@@ -6,24 +7,24 @@ pub struct Market {
     pub name: String,
     pub base_currency: Option<String>,
     pub quote_currency: Option<String>,
-    pub quote_volume_24h: Option<f64>,
-    pub change_1h: Option<f64>,
-    pub change_24h: Option<f64>,
+    pub quote_volume_24h: Option<Decimal>,
+    pub change_1h: Option<Decimal>,
+    pub change_24h: Option<Decimal>,
     pub high_leverage_fee_exempt: bool,
-    pub min_provide_size: Option<f64>,
+    pub min_provide_size: Option<Decimal>,
     pub r#type: MarketType,
     pub underlying: Option<String>,
     pub enabled: bool,
-    pub ask: Option<f64>,
-    pub bid: Option<f64>,
-    pub last: Option<f64>,
+    pub ask: Option<Decimal>,
+    pub bid: Option<Decimal>,
+    pub last: Option<Decimal>,
     pub post_only: bool,
-    pub price: Option<f64>,
-    pub price_increment: Option<f64>,
-    pub size_increment: Option<f64>,
+    pub price: Option<Decimal>,
+    pub price_increment: Option<Decimal>,
+    pub size_increment: Option<Decimal>,
     pub restricted: bool,
-    pub volume_usd_24h: Option<f64>,
-    pub large_order_threshold: Option<f64>,
+    pub volume_usd_24h: Option<Decimal>,
+    pub large_order_threshold: Option<Decimal>,
     pub is_etf_market: bool,
 }
