+++ src/rest/model/orderbook.rs
@@ -1,5 +1,7 @@
+use rust_decimal::Decimal;
+
 #[derive(Debug, serde::Serialize, serde::Deserialize)]
 pub struct Orderbook {
-    pub asks: Vec<[f64; 2]>,
-    pub bids: Vec<[f64; 2]>,
+    pub asks: Vec<[Decimal; 2]>,
+    pub bids: Vec<[Decimal; 2]>,
 }
