+++ src/rest/model/transfer.rs
@@ -1,8 +1,10 @@
+use rust_decimal::Decimal;
+
 #[derive(Debug, serde::Serialize, serde::Deserialize)]
 pub struct Transfer {
     pub id: u64,
     pub coin: String,
-    pub size: f64,
+    pub size: Decimal,
     pub time: chrono::DateTime<chrono::Utc>,
     pub notes: String,
     pub status: String,
